+++ CHANGES.md
@@ -39,3 +39,4 @@
 * added authentication required to encrypt the private key with the user's password
 * updated Ubuntu deployment scripts
 * derive pub key from cert; priv key encrypted with user password, not available on smtp relay
+* multiple ports connection option for SMTP server